Allow customizing delay between DFU connections
All DFU implementations currently have had a 3 second delay before
connecting to the DFU target device. This delay is needed because the
target needs some time to reboot after sending images.

There are, however, cases where this delay is too low or too high.
Instead of keeping this as a hard-coded value, we now allow the user to
pass a `--connect-delay <seconds>` parameter. If not specified, then it
will default to 3 seconds.