Add info about MSD in certain error scenarios
When running serial DFU on Ubuntu 16.04 with PCA10040 with MSD enabled,
communication errors may occur. Adding info about MSD and how to disable
it as part of the error messages.

There may be more scenarios where serial DFU can fail because of MSD
being enabled, but these were the ones I was able to provoke.