Distinguish serial port timeout and ping timeout
and add option to set them. By default the update now wait 30 seconds
before timeout because of a no ping response.
This enable user to run the update utility before going in dfu state as
the timing can be quite short with some DFU settings. For instance, if
the only way to trigger the update is during the boot process after a
reset with a short inactivity timer.