Add options to configure serial update timeouts (#199)
* Don't sleep while between device ping when doing UART serial update as we already wait in read operation.

* Stop pinging target when it responds while performing UART serial update as it already respond, so there is no need to continue to send ping.

* Distinguish serial port timeout and ping timeout

and add option to set them. By default the update now wait 30 seconds
before timeout because of a no ping response.
This enable user to run the update utility before going in dfu state as
the timing can be quite short with some DFU settings. For instance, if
the only way to trigger the update is during the boot process after a
reset with a short inactivity timer.