Port to Python 3 (#250)
* Initial py3 support

This is work in progress

* Use intelhex pkg, fix integer division

* Use intelhex module

* Update pyyaml to latest version

* Fix typo

* Update module reqs

* ecdsa module do not support bytearray, use bytes

* Fix binary write to file and hex conversion

* Use intelhex pkg

* Use bytes and not decode

* Convert hci.slip to Python 3

* Py3ize part of bl_dfu_sett

* Fix assertion to expect bytes

* Gitignore tests run artefacts

* Fix str used as bytes

* Py3ize Signing

There is a rather arbitrary decision make in this change that PEMs are
represented in bytes. It was selected in the interest of keeping
conversions to a minimum. Both bytes and str are good candidated for
representing PEM, which consists only of ascii chars.

* Upgrade pc_ble_driver_py to 0.12 for py3 support

* Py3ize zigbee/prod_config

* Remove unneeded NoneType import that breaks py3

Python's types.NoneType is no longer available, and it was not used
anyway.

* Add None guard before comparing as needed for py3

* Add some py3 support in TestDfuTransportSerial

The test will still not complete due to changes in interal apis.

* Adjust requirements for py3

* Evaluate map object of serial ports to list

* Py3ize dfu_transport_ble

* Port dfu_transport_ble to SDv5

The new pc_ble_driver_py module with py3 support is SDv5 based.

* Change type warning string to bytes

The alternative is to instead .decode('ascii') the result from get_sk,
but since everyting is ascii, bytes is a suitable (as intended) type
that can actually be safer in case PEM files cannot contain non-ascii
values.

* Remove map(ord, ...) on struct.pack in Ant

struct.pack in Python 3 returns bytes which gives ints when iterated
over.

* Update requirements.txt and add requirements-dev.txt

* Change readme to run package instead of __main__.py

This way, importing nordicsemi works without changes to PYTHONPATH.

* Remove Python 2 specific note in README

* Add tests artefacts to gitignore

* Remove vestigal debug print from test

* Use ABC instead of ABCMeta

* Fixup copyright year to a range