Remove dependency version constraints
Remove most dependency version constraints from requirements.txt, which
is read by setup.py. This was done so that any project that depends on
this one can choose dependency versions as freely as possible and is
reposible for testing and freezing for its releases. This change also
introduces a requirements-freeze.txt which freezes the dependency
environment to a known good state for nrfutil that will be used for
builds of nrfutil standalone executable.

Sadly this means that installing nrfutil manually trough pip can break
without warning. A possible solution for this is to create a dedicated
package for nrfutil with the frozen dependencies. This would be ment to
be run in a isolated environment using pipx or a simmilar tool.