Revert change of antlib version back to 1.1b0.post0
It is unclear what is going on with antlib releases. PyPI.org show
antlib has releases 1.0b0 and 1.1b0 but no 1.1b.post0 (a versioning
scheme discouraged by PEP440), but the only wheel avaliable for download
is for 1.1b.post0. It seems pip is looking only at the files available
for download, so installing 1.1b0.post0 works for now.