dfu_trigger: fix handling of lower than 0x1000 VIDs and PIDs
VID and PID are passed as strings to the select_device(), but getVendorID()
and getProductID() are returning with integers. Currently the integers are
converted to hex strings to compare those with the IDs we are searching for.
But with the current int-to-string conversion, the leading 0s are truncated,
so such IDs will never match.

Signed-off-by: Robert Hodaszi <hodrob84@gmail.com>