Align dfu zigbee command with pc-ble-driver>=0.15.0 (#340)
The API of call_cmd method changed in pc-ble-driver-py 0.15.0. Previously, the CalledProcessError was raised when nrfjprog command failed. Since version 0.15.0, it is replaced by RuntimeError.
Because RuntimeError has no returncode attribute, there is no good way to check the reason why command failed (parsing command output is not recommended, cause it can be changed).
When requirements of nrfutil are updated, the old code can be removed.